separates reducers
