+++ src/reducers/index.js
@@ -1,16 +1,7 @@
-import { LOAD_CARDS, ADD_CARD } from '../actions/cards';
+import { combineReducers } from 'redux';
 
-const initialState = [ ];
+import cards from './cards';
 
-const reducers = (state = initialState, action) => {
-  switch(action.type){
-    case LOAD_CARDS:
-      return [ ...action.cards ];
-    case ADD_CARD:
-      return [ ...state, action.card ];
-    default:
-      return state
-  }
-}
-
-export default reducers;
+export default combineReducers({
+  cards
+});
